+++ red-bacnet/discover_point.js
@@ -3,9 +3,10 @@ require('./_alias.js');
 
 const EventEmitter = require('events');
 
+const { print } = require('@root/common/core/util.js')
 const { nowFormatted } = require('@root/common/core/util.js')
 const { DiscoverPointJob } = require('@root/common/job/discover_point.js')
-const { CoalescedJobQueue } = require('@root/common/job/core.js')
+const { getQueue } = require('@root/common/job/global_queue_manager.js')
 const {
     EVENT_UPDATE_STATUS, EVENT_ERROR, EVENT_INPUT, EVENT_OUTPUT
 } = require('@root/common/core/constant.js')
@@ -30,9 +31,8 @@ module.exports = function (RED) {
             // events
             this.#subscribeListeners();
 
-            // configure job queue
-            this.job = new CoalescedJobQueue();
-            this.job.run();
+            // configure job queue - use global queue
+            this.job = getQueue();
         }
 
         #subscribeListeners() {
@@ -41,6 +41,14 @@ module.exports = function (RED) {
             */
             // @ts-ignore
             this.on(EVENT_INPUT, async function (msg) {
+                const jobId = (typeof msg.id === 'string' || typeof msg.id === 'number') ? msg.id : 'discoverPoints';
+
+                if (this.job.queue.map(item => item.id).includes(jobId)) {
+                    // @ts-ignore
+                    print(`Coalesced job: ${jobId}`, true)
+                    return;
+                }
+
                 const task = new DiscoverPointJob(
                     this.client, this.#eventEmitter, msg?.devices, this.discoverMode, this.readMethod,
                     this.groupExportDeviceCount, this.maxConcurrentDeviceRead,
@@ -48,9 +56,12 @@ module.exports = function (RED) {
                 );
 
                 this.job.addJob({
-                    id: (typeof msg.id === 'string' || typeof msg.id === 'number') ? msg.id : 'task',
-                    task: task
+                    id: jobId,
+                    task: task,
+                    priority: Number.isFinite(msg.priority) ? msg.priority : 5,
                 });
+                // @ts-ignore
+                this.status({ fill: 'yellow', shape: 'dot', text: `in queue` });
             });
 
             this.#eventEmitter.on(EVENT_OUTPUT, (data) => {